Conflicts:
	Atmosphere/LibImport.txt
	Atmosphere/MenuBME.h
	Atmosphere/MenuWelcome.cpp
	Atmosphere/MenuWelcome.h
	Atmosphere/main.cpp
	Library/LibESP8266/ATClient.cpp
	Library/LibESP8266/ATClient.h
	Library/LibESP8266/ESP8266Wifi.cpp
	Library/LibESP8266/ESP8266Wifi.h
	Library/LibHTTP/HTTPRequest.cpp
	Library/LibHTTP/HTTPRequest.h
	Library/LibHTTP/HTTPServer.cpp
	Library/LibHTTP/HTTPServer.h
	Library/LibLCD/LCDButton.cpp
	Library/LibLCD/LCDDisplay.cpp
	Library/LibLCD/LCDDisplay.h
	Library/LibMemory/Buffer.cpp
	Library/LibMemory/Buffer.h
	Library/LibMenus/SubMenu.cpp
	Library/LibNetwork/NetworkStream.cpp
	Library/LibNetwork/NetworkStream.h
	Moisture/AutomaticWater.cpp
	Moisture/AutomaticWater.h
	Moisture/LCDWaterDisplay.cpp
	Moisture/LCDWaterDisplay.h
	Moisture/MoistureSensor.cpp
	Moisture/MoistureSensor.h
	Moisture/PumpControl.cpp
	Moisture/PumpControl.h
	Moisture/ValveController.cpp
	Moisture/ValveController.h
	Moisture/main.cpp
	Wifi/main.cpp

+++ Moisture/MoistureSensor.h
@@ -51,7 +51,7 @@ private:
   unsigned int  _tickInterval;    /** Length of a tick [ms] */
   unsigned int  _currentCounter;  /** Internal tick counter */
 
-  unsigned short _currentValueIndex; /** Index of the current value in the _valArray buffer */
+  uint8_t _currentValueIndex;        /** Index of the current value in the _valArray buffer */
   int _lastValue;                    /** Last raw value read */
   int _valArray[30];                 /** Buffer of values used for averaging during calibration */
   int _WATER_VALUE;                  /** Calibration value for maximum moisture */
